Refactor: Flatten module hierarchy for errors, codegen, and grammar
Flattened `src/errors/`, `src/codegen/`, and `src/grammar/` directories into single files to reduce complexity and file count.
- `src/errors/` -> `src/errors.rs`
- `src/codegen/` -> `src/codegen.rs`
- `src/grammar/` -> `src/grammar.rs`
- Moved `src/grammar/glossa.pest` to `src/glossa.pest`

Adheres to Razor's KISS principle by removing unnecessary nesting.
