+++ tests/assembly_error_coverage.rs
@@ -1,5 +1,5 @@
 use glossa::errors::GlossaError;
-use glossa::errors::assembly::AssemblyError;
+use glossa::errors::AssemblyError;
 use glossa::morphology::{Case, Gender, Number, Person, analyze};
 use glossa::semantic::assembler::Assembler;
